% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-helpers.R
\name{collapse_vec}
\alias{collapse_vec}
\title{Collapse Vector to Sentence}
\usage{
collapse_vec(x)
}
\arguments{
\item{x}{Strings to collapse.}
}
\value{
String with appropriate commas/and.
}
\description{
Collapse vector of strings to formal sentence.
}
