% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-helpers.R
\name{df_arsenal}
\alias{df_arsenal}
\title{Arsenal Data Frame}
\usage{
df_arsenal(x, labs = NULL, ...)
}
\arguments{
\item{x}{arsenal::tableby object.}

\item{labs}{Label translations.}

\item{...}{Additional arguments sent to arsenal::summary.tableby.}
}
\value{
knitr::kable output of data preview.
}
\description{
Convert arsenal::tableby to data frame.
}
