% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_mec.R
\name{filter_mec}
\alias{filter_mec}
\title{Filter MEC Data Objects}
\usage{
filter_mec(x, study)
}
\arguments{
\item{x}{The MEC data object to be filtered.}

\item{study}{The MEC study to filter to.}
}
\value{
The filtered data object
}
\description{
Filters MEC data objects to pre-defined sample sets
}
