% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-helpers.R
\name{formadoo}
\alias{formadoo}
\title{Create Formula}
\usage{
formadoo(x, adj = NULL, y = NULL, ...)
}
\arguments{
\item{x}{variable of interest}

\item{adj}{adjustment variables}

\item{y}{response variable}

\item{...}{Arguments passed to reformulate()}
}
\value{
formula
}
\description{
formula with variable of interest placed as the last term
}
