% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-helpers.R
\name{get_paov}
\alias{get_paov}
\title{Get Permanova}
\usage{
get_paov(df, x, adj, fm, uni, md = c("a1", "a2"), np = 2,
  seed = 123456, ...)
}
\arguments{
\item{df}{source data.frame}

\item{x}{variable of interest}

\item{adj}{adjustment varaiables}

\item{fm}{MEC filter}

\item{uni}{unifrac_uw or unifrac_wt}

\item{md}{a1 = adonis() or a2 = adonis2()}

\item{np}{number of permutations}

\item{seed}{seed for reproducibility}

\item{...}{arguments passed to adonis()/adonis2()}
}
\value{
list with unifrac label and model results
}
\description{
Run custon adonis()/adonis2()
}
