% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-helpers.R
\name{get_paov_data}
\alias{get_paov_data}
\title{Get Permanova Data}
\usage{
get_paov_data(df, x, adj, fm, uni = c("unifrac_uw", "unifrac_wt"))
}
\arguments{
\item{df}{source data.frame}

\item{x}{variable of interest}

\item{adj}{adjustment varaiables}

\item{fm}{MEC filter}

\item{uni}{unifrac_uw or unifrac_wt}
}
\value{
list with unifrac distance object and predictor data.frame
}
\description{
Retrieve and polisth Unifrac matrix and predictor data
}
