% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-helpers.R
\name{get_polr}
\alias{get_polr}
\title{Get polr()}
\usage{
get_polr(df, x, adj, fm, bug = c("phyla", "genera"), ...)
}
\arguments{
\item{df}{source data.frame}

\item{x}{variable of interest}

\item{adj}{adjustment varaiables}

\item{fm}{MEC filter}

\item{bug}{genera or phyla}

\item{...}{Arguments passed to polr()}
}
\value{
list with tidy(), glance(), and lrtest() results
}
\description{
Run custom polr()
}
