% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-helpers.R
\name{get_polr_data}
\alias{get_polr_data}
\title{Get Data for polr()}
\usage{
get_polr_data(df, x, adj, fm, bug = c("phyla", "genera"), ...)
}
\arguments{
\item{df}{source data.frame}

\item{x}{variable of interest}

\item{adj}{adjustment varaiables}

\item{fm}{MEC filter}

\item{bug}{genera or phyla}

\item{...}{Arguments passed to proc_rank()}
}
\value{
nested tibble
}
\description{
Retrieve, polish and quintile data for polr() bug models
}
