% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-helpers.R
\name{get_polr_result}
\alias{get_polr_result}
\title{Get Results for polr()}
\usage{
get_polr_result(m1, lrt)
}
\arguments{
\item{m1}{polr() object}

\item{lrt}{lrtest() object}
}
\value{
list with tidy(), glance(), and lrtest() results
}
\description{
Create list of polr() and lrtest() results
}
