% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-helpers.R
\name{get_tableby}
\alias{get_tableby}
\title{Get tableby()}
\usage{
get_tableby(df, y, x, fm, lab = NULL, ...)
}
\arguments{
\item{df}{source data.frame}

\item{y}{response variable}

\item{x}{variable of interest}

\item{fm}{MEC filter}

\item{lab}{Label for x}

\item{...}{Arguments passed to tableby()}
}
\value{
tableby object
}
\description{
Run custom tableby()
}
