% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-helpers.R
\name{gsub_geq}
\alias{gsub_geq}
\title{Replace ">="}
\usage{
gsub_geq(x)
}
\arguments{
\item{x}{String to modify.}
}
\value{
String with unicode symbol for ">="
}
\description{
Uses `gsub()` to replace ">=" to the unicode symbol.
}
