% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-helpers.R
\name{make_dl_link}
\alias{make_dl_link}
\title{Link to Data}
\usage{
make_dl_link(x, path, timestamp = TRUE)
}
\arguments{
\item{x}{Data frame to be linked to.}

\item{path}{File path to write the downloadable data to.}

\item{timestamp}{Add a timestamp to the file name.}
}
\value{
HTML code for link

File saved to disk for download
}
\description{
Makes a hyperlink to a downloadable data object.
}
