% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-helpers.R
\name{mec_field}
\alias{mec_field}
\title{Return pretty field name}
\usage{
mec_field(x)
}
\arguments{
\item{x}{Field to format}
}
\value{
The pretty field name
}
\description{
Returns the formatted name of various MEC data fields.
}
