% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-helpers.R
\name{polish}
\alias{polish}
\alias{polish.data.frame}
\alias{polish.matrix}
\alias{polish.dist}
\title{Polish Data Objects}
\usage{
polish(x, ...)

\method{polish}{data.frame}(x, ...)

\method{polish}{matrix}(x, cols = NULL, rows = NULL, ...)

\method{polish}{dist}(x, id = NULL, mat = FALSE, ...)
}
\arguments{
\item{x}{The object to be polished}

\item{...}{Columns passed to dplyr::select}

\item{cols}{Columns to keep from matrix}

\item{rows}{Rows to keep from matrix}

\item{id}{Names/index to keep from dist}

\item{mat}{Return dist object as matrix}
}
\value{
data.frame with specified colums or matrix/dist with specified rows/columns
}
\description{
Select specified variables from data.frame or rows/columns from matrix/dist
}
