% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-helpers.R
\name{preview_tbl}
\alias{preview_tbl}
\title{Preview Tibble}
\usage{
preview_tbl(x, cols = 5, rows = 10, ...)
}
\arguments{
\item{x}{Data to preview.}

\item{cols}{Number of column to preview.}

\item{rows}{Number of rows to preview.}

\item{...}{Additional arguments sent to knitr::kable}
}
\value{
knitr::kable output of data preview.
}
\description{
Shows a preview of a tibble/data.frame.
}
