% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-helpers.R
\name{proc_rank}
\alias{proc_rank}
\title{SAS-like Proc Rank() for binning}
\usage{
proc_rank(x, k = 5, ties = "min")
}
\arguments{
\item{x}{numeric vector}

\item{k}{number of bins}

\item{ties}{how to break ties (see rank)}
}
\value{
ordered factor
}
\description{
Run SAS-like proc rank to create bins
}
