% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-helpers.R
\name{retrieve}
\alias{retrieve}
\alias{retrieve.default}
\alias{retrieve.list}
\title{Retrieve Data Objects}
\usage{
retrieve(x, ...)

\method{retrieve}{default}(x, ..., path = "Data")

\method{retrieve}{list}(x, ..., path = "Data")
}
\arguments{
\item{x}{Data to be loaded}

\item{...}{Arguments passed to polish()}

\item{path}{Path to data directory}
}
\value{
Polished data.frame, matrix, or dist
}
\description{
Load data (.rds file) and polish
}
