% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-helpers.R
\name{sniff}
\alias{sniff}
\alias{sniff.default}
\alias{sniff.hurdle}
\alias{sniff.adonis}
\title{Sniff Model Objects}
\usage{
sniff(x, ...)

\method{sniff}{default}(x, ...)

\method{sniff}{hurdle}(x, ...)

\method{sniff}{adonis}(x, ...)
}
\arguments{
\item{x}{The model to be sniffed.}

\item{...}{Additional arguments sent to `broom::tidy`.}
}
\value{
Tibble containing the sniffed out model info
}
\description{
Sniff out and return model information
}
