% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study-parameters.R
\name{study_ancestry}
\alias{study_ancestry}
\title{Genetic Ancestry Study Parameters}
\usage{
study_ancestry(main_var = "anc_max", main_var_cols = c("#A6CEE3",
  "#1F78B4", "#B2DF8A", "#33A02C", "#FB9A99", "#E31A1C", "#FF7F00",
  "#6A3D9A"), main_cont = "", dir_out = "Output/Ancestry",
  dir_dta = file.path(dir_out, "Data"), dir_dwn = file.path(dir_out,
  "Downloads"), dir_fig = file.path(dir_out, "Figures"),
  dir_tab = file.path(dir_out, "Tables"), study_filter = "mGWAS",
  study_title = "Ancestry", sg = "ancestry", q_cut = c(phylum = 0.05,
  genera = 0.05), zhern_cols = viridis::viridis(10),
  zhern_names = c("Demographics", "Alcohol", "Anthropometric measures",
  "Special diet", "Cultural foods & dietary components",
  "Dietary quality index", "Medical conditions", "Medications",
  "Physical activity", "Smoking"), run_code = FALSE,
  download_files = TRUE)
}
\arguments{
\item{main_var}{Main study categorical variable}

\item{main_var_cols}{Colors for main study varible}

\item{main_cont}{Main study continuous variable}

\item{dir_out}{Output directory}

\item{dir_dta}{Output directory for data objects}

\item{dir_dwn}{Output directory for downloadable files}

\item{dir_fig}{Output directory for figures}

\item{dir_tab}{Output directory for tables}

\item{study_filter}{Study indicator for filtering to appropriate samples}

\item{study_title}{Formal study title}

\item{sg}{Lowercase study group name}

\item{q_cut}{FDR cutoffs for phylum and genera}

\item{zhern_cols}{Colors for Zhernakova categories}

\item{zhern_names}{Names of Zhernakova categories}

\item{run_code}{Run scripts during book build}

\item{download_files}{Create downloadable file links}
}
\value{
Variables are saved to the global environment.
}
\description{
Set global variables for the MEC Characterization paper
}
