% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study-parameters.R
\name{study_pancreas_fat}
\alias{study_pancreas_fat}
\title{Pancreas Fat Study Parameters}
\usage{
study_pancreas_fat(main_var = "OMRI_NAFPD_55",
  main_var_cols = c("dodgerblue", "red"),
  main_cont = "OMRI_pct_panc_fat_corr",
  dir_out = "Output/Pancreas-Fat", dir_dta = file.path(dir_out,
  "Data"), dir_dwn = file.path(dir_out, "Downloads"),
  dir_fig = file.path(dir_out, "Figures"), dir_tab = file.path(dir_out,
  "Tables"), study_filter = "PF", study_title = "Pancreas Fat",
  study_sent = "Pancreas fat", sg = "pancreas fat", q_cut = c(phylum
  = 0.2, genera = 0.05), cf = "_bookdown-pancreas-fat.yml",
  run_code = FALSE, download_files = TRUE)
}
\arguments{
\item{main_var}{Main study categorical variable}

\item{main_var_cols}{Colors for main study varible}

\item{main_cont}{Main study continuous variable}

\item{dir_out}{Output directory}

\item{dir_dta}{Output directory for data objects}

\item{dir_dwn}{Output directory for downloadable files}

\item{dir_fig}{Output directory for figures}

\item{dir_tab}{Output directory for tables}

\item{study_filter}{Study indicator for filtering to appropriate samples}

\item{study_title}{Formal study title}

\item{study_sent}{Study title for sentence}

\item{sg}{Lowercase study group name}

\item{q_cut}{FDR cutoffs for phylum and genera}

\item{cf}{bookdown config file}

\item{run_code}{Run scripts during book build}

\item{download_files}{Create downloadable file links}
}
\value{
Variables are saved to the global environment.
}
\description{
Set global variables for the MEC Pancreas Fat paper
}
