% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-helpers.R
\name{tbl_strata}
\alias{tbl_strata}
\title{Stratify Data}
\usage{
tbl_strata(df, strata, name_all = "All", sep = "___")
}
\arguments{
\item{df}{The data to be stratified.}

\item{strata}{List of vectors to stratify by.}

\item{name_all}{The indicator name for the strata containing all observations}

\item{sep}{Separator}

\item{...}{Additional arguments sent to `broom::tidy`.}
}
\value{
Tibble containing stratified data.

strata should be a list of vectors to stratify df by;
each vector should have length = nrow(df);
if named list, then the names become
columns in the data frame (duplicate names are ajusted);
if not a named list, then generic names are made
}
\description{
Stratify a data frame by a list of strata
}
