% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-helpers.R
\name{tidy.hurdle}
\alias{tidy.hurdle}
\title{Tidy Hurdle Model}
\usage{
tidy.hurdle(hm, exponentiate = FALSE, ...)
}
\arguments{
\item{hm}{The hurdle model.}

\item{exponentiate}{Exponentiate the estimates.}

\item{...}{Additional arguments passed to `confint`.}
}
\value{
Tibble containing the tidied hurdle model.
}
\description{
broom::tidy.hurdle method
}
