% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-objects.R
\name{filter_mec}
\alias{filter_mec}
\alias{filter_mec.default}
\alias{filter_mec.data.frame}
\alias{filter_mec.matrix}
\alias{filter_mec.dist}
\title{Filter MEC Data Objects}
\usage{
filter_mec(x, study)

\method{filter_mec}{default}(x, study = c("1", "2", "mGWAS", "MEC-APS",
  "LF5", "LF55", "PF5", "PF55", "VF", "FHC", "Other"))

\method{filter_mec}{data.frame}(x, study)

\method{filter_mec}{matrix}(x, study)

\method{filter_mec}{dist}(x, study)
}
\arguments{
\item{x}{The MEC data object to be filtered.}

\item{study}{The MEC study to filter to.}
}
\value{
The filtered data object
}
\description{
Filters MEC data objects to pre-defined sample sets
}
